Added support for polymeric audit trail
Some times a project contains more than one state_machine model.
This modification allows to manage all the auditing of multiple models
into one single table.
This commit contains the logic to store the transitions into one single
table.
A modification to the generator to create the single table model.
-TODO:
Adapt the tests to work with this approach.